+++ yAppWeb/ReleaseNotes.md
@@ -6,7 +6,9 @@
 * devTaskIssue#48-FeatureIssue#1-Create post object
 * devTaskIssue#45-FeatureIssue#1-Setup a post table schema using C#
 * devTaskIssue#49-FeatureIssue#1-Issue #7 unit tests
+* devTaskIssue#50-FeatureIssue#1-Issue #7 integration tests
 * devTaskIssue#63-FeatureIssue#1-Create an endpoint to get all public question posts a user has made
+
 * devTaskIssue#38-FeatureIssue#5-Create an endpoint to allow user to update their information
 * devTaskIssue#39-FeatureIssue#5-Create an endpoint to get all friends
 * devTaskIssue#40-FeatureIssue#5-Setup a friend table schema using c# #40